+++ src/ChatControl.php
@@ -11,7 +11,7 @@ namespace HeroesofAbenez\Chat;
  * @property-read \Nette\Bridges\ApplicationLatte\Template $template
  */
 abstract class ChatControl extends \Nette\Application\UI\Control {
-  /** @var IChatMessageProcessor[] */
+  /** @var ChatMessageProcessor[] */
   protected array $messageProcessors = [];
   protected string $characterColumn;
   /** @var int */
@@ -20,7 +20,7 @@ abstract class ChatControl extends \Nette\Application\UI\Control {
   protected string $templateFile = __DIR__ . "/chat.latte";
   protected int $messagesPerPage = 25;
 
-  public function __construct(protected IDatabaseAdapter $database, protected string $textColumn, protected int $textValue, string $characterColumn = null, mixed $characterValue = null) {
+  public function __construct(protected DatabaseAdapter $database, protected string $textColumn, protected int $textValue, string $characterColumn = null, mixed $characterValue = null) {
     $this->characterColumn = $characterColumn ?? $textColumn;
     $this->characterValue = $characterValue ?? $textValue;
   }
@@ -56,7 +56,7 @@ abstract class ChatControl extends \Nette\Application\UI\Control {
     $this->characterProfileLink = $characterProfileLink;
   }
   
-  public function addMessageProcessor(IChatMessageProcessor $processor): void {
+  public function addMessageProcessor(ChatMessageProcessor $processor): void {
     $this->messageProcessors[] = $processor;
   }