+++ src/IDatabaseAdapter.php
@@ -3,20 +3,11 @@ declare(strict_types=1);
 
 namespace HeroesofAbenez\Chat;
 
-/**
- * IDatabaseAdapter
- *
- * @author Jakub Konečný
- */
-interface IDatabaseAdapter {
-  /**
-   * @param mixed $value
-   */
-  public function getTexts(string $column, $value, int $limit): ChatMessagesCollection;
-  /**
-   * @param mixed $value
-   */
-  public function getCharacters(string $column, $value): ChatCharactersCollection;
-  public function addMessage(string $message, string $filterColumn, int $filterValue): void;
+if(false) {
+  /** @deprecated use DatabaseAdapter */
+  interface IDatabaseAdapter extends DatabaseAdapter {
+  }
+} else {
+  class_alias(DatabaseAdapter::class, IDatabaseAdapter::class);
 }
 ?>