+++ tests/HeroesofAbenez/Chat/Test2Command.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace HeroesofAbenez\Chat;
 
-final class Test2Command extends ChatCommand {
+final class Test2Command extends BaseChatCommand {
   public const NAME = "test2";
   
   public function execute() : string {