+++ tests/HeroesofAbenez/Chat/TestCommand.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace HeroesofAbenez\Chat;
 
-final class TestCommand extends ChatCommand {
+final class TestCommand extends BaseChatCommand {
   /** @var string */
   protected string $name = "test1";