+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 2.2.0-dev
 - raised minimal version of PHP to 8.1
 - deprecated most getters and setters
+- deprecated IChatMessageProcessor in favor of ChatMessageProcessor, IDatabaseAdapter in favor of DatabaseAdapter and ChatCommand in favor of BaseChatCommand
 
 Version 2.1.0
 - raised minimal version of PHP to 7.4